refactor(renderer): publish the module's dependency free leaves
