+++ src/modules/settings/themeOptions.ts
@@ -1,4 +1,4 @@
 // The registry itself lives in the renderer, which is what defines these settings. This file keeps
 // the extension-side import path stable until the call sites move into the module.
 
-export { registerThemeSetting, setThemeSettings } from "@renderer/index";
+export { registerThemeSetting, setThemeSettings } from "@renderer/themeSettings";
